% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoPolygons.R
\name{add-tgm-polygons}
\alias{add-tgm-polygons}
\alias{addTargomoPolygons}
\title{Add Targomo Polygons}
\usage{
addTargomoPolygons(map, api_key = Sys.getenv("TARGOMO_API_KEY"),
  lat = 55.9, lng = -3.1, options = targomoPolygonOptions(),
  fitBounds = TRUE)
}
\arguments{
\item{map}{A leaflet map}

\item{api_key}{Your Targomo API key - recommend to use the 'TARGOMO_API_KEY' environment variable}

\item{lat, lng}{Coordinates of the source of the polygons}

\item{options}{See targomoOptions}

\item{fitBounds}{Should the map rezoom to fit the bounds of the polygons?}
}
\description{
This function and the associated options allow you to easily add isochrones from
the Targomo Polygon API to your leaflet map.
}
