% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoPolygons.R
\name{addTargomoPolygons}
\alias{addTargomoPolygons}
\title{Add Targomo Polygons}
\usage{
addTargomoPolygons(map, api_key = Sys.getenv("TARGOMO_API_KEY"),
  lng = 0, lat = 51, times = c(600, 1200, 1800),
  transport = "walk", stroke = 10)
}
\description{
Add Targomo Polygons
}
