% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoRoutes.R
\name{getTargomoRoutes}
\alias{getTargomoRoutes}
\title{Get Targomo Routes}
\usage{
getTargomoRoutes(api_key = Sys.getenv("TARGOMO_API_KEY"),
  region = Sys.getenv("TARGOMO_REGION"), source_data = NULL,
  source_lat = NULL, source_lng = NULL, target_data = NULL,
  target_lat = NULL, target_lng = NULL, options = targomoOptions(),
  verbose = FALSE, progress = FALSE)
}
\description{
Get Targomo Routes
}
