% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\name{addTimeLegend}
\alias{addTimeLegend}
\title{Add Time Legend to Map}
\usage{
addTimeLegend(map, palette, values, options, group)
}
\arguments{
\item{map}{A leaflet map}

\item{palette}{A colour palette (from \code{\link{createTimePalette}})}

\item{values}{Values to use (travel times)}

\item{options}{A set of \code{\link{timeLegendOptions}}}

\item{group}{The layer group to add the legend to}
}
\description{
Add Time Legend to Map
}
