% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoDependency.R
\name{api-versions}
\alias{api-versions}
\alias{targomoCoreVersion}
\alias{targomoLeafletVersion}
\title{Targomo Version}
\usage{
targomoCoreVersion()

targomoLeafletVersion()
}
\description{
These functions return the versions of the Targomo API which are bundled with
the package
}
