% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{attribution}
\alias{attribution}
\alias{attributionFreeIframe}
\alias{attributionOtherIframe}
\alias{attributionLink}
\alias{addTargomoAttribution}
\title{Add Targomo Attributions}
\usage{
attributionFreeIframe()

attributionOtherIframe()

attributionLink()

addTargomoAttribution(map, free_plan = TRUE, ...)
}
\arguments{
\item{map}{A leaflet map}

\item{free_plan}{Logical - is the Targomo plan you're using free or paid?}

\item{...}{Further arguments to pass to \code{leaflet::addControl} e.g. position}
}
\description{
Functions providing link to Targomo Attributions page, depending on plan.
}
