% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{basemaps}
\alias{basemaps}
\alias{targomoMapStyles}
\alias{getTargomoMapURL}
\alias{addTargomoTiles}
\title{Add Targomo Basemaps to a Leaflet Map}
\usage{
targomoMapStyles()

getTargomoMapURL(style = "basic", api_key = NULL)

addTargomoTiles(map, style = "basic",
  api_key = Sys.getenv("TARGOMO_API_KEY"), layerId = NULL,
  group = NULL, ...)
}
\arguments{
\item{style}{A valid Targomo Map Style - see \code{targomoMapStyles()}}

\item{api_key}{Your Targomo API key - defaults to the \code{TARGOMO_API_KEY}
ennvironment variable}

\item{map}{A leaflet map}

\item{layerId}{The layer id to pass to \code{leaflet::addTiles}}

\item{group}{The layer group to pass to \code{leaflet::addTiles}}

\item{...}{Further options to pass to \code{leaflet::addTiles} e.g. options}
}
\description{
This function wraps round \code{leaflet::addTiles} to provide access to the Targomo basemaps.
}
