% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon-helpers.R
\name{callPolygonService}
\alias{callPolygonService}
\title{Call the Targomo Polygon Service}
\usage{
callPolygonService(api_key, region, sources, options, verbose = FALSE,
  progress = FALSE)
}
\arguments{
\item{api_key}{The Targomo API key.}

\item{region}{The Targomo region.}

\item{sources}{A processed sources object to pass to the API.}

\item{options}{A processed options list.}

\item{verbose}{Display info on the API call?}

\item{progress}{Display a progress bar?}
}
\description{
Call the Targomo Polygon Service
}
