% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn-polygon.R
\name{createPolygonOptions}
\alias{createPolygonOptions}
\title{Create options}
\usage{
createPolygonOptions(options)
}
\arguments{
\item{options}{The output of \code{targomoOptions}.}
}
\description{
Function to create options in a nested list structure suitable to be turned into JSON.
}
