% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn-polygon.R
\name{createPolygonRequestBody}
\alias{createPolygonRequestBody}
\title{Create Request Body}
\usage{
createPolygonRequestBody(sources, options)
}
\arguments{
\item{sources}{A processed sources object to pass to the API.}

\item{options}{A processed options list.}
}
\description{
Function to create a request body using the sources and options given.
}
