% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon-helpers.R
\name{createPolygonSources}
\alias{createPolygonSources}
\title{Create Sources}
\usage{
createPolygonSources(data, lat, lng, options)
}
\arguments{
\item{data}{The data object}

\item{lat, lng}{The lat/lng vectors or formulae to resolve}

\item{options}{A processed options object.}
}
\description{
Function to create the sources needed to query the Targomo API.
}
