% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoPolygons.R
\name{createRequestBody}
\alias{createRequestBody}
\title{Create Request Body}
\usage{
createRequestBody(sources, options)
}
\description{
Function to create a request body using the sources and options given
}
