% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{createRequestURL}
\alias{createRequestURL}
\title{Create Request URL}
\usage{
createRequestURL(region, end_point)
}
\arguments{
\item{region}{The Targomo region.}

\item{end_point}{The API end_point.}
}
\value{
Character string, the URL of the chosen endpoint
}
\description{
Function to create the request URL.
}
