% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\name{createTimePalette}
\alias{createTimePalette}
\title{Create a Colour Palette for Time Service Results}
\usage{
createTimePalette(palette, type, maxTime, bins, reverse)
}
\arguments{
\item{palette}{A colour palette e.g. "viridis", "Blues"}

\item{type}{Either "numeric" or "bin"}

\item{maxTime}{The maximum time value to consider}

\item{bins}{Either a single number of bins, or a vector of cut points.}

\item{reverse}{Whether to reverse the colour palette.}
}
\value{
A colour palette function for use with the time legend and markers
}
\description{
Create a Colour Palette for Time Service Results
}
