% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoDependency.R
\name{dependencies}
\alias{dependencies}
\alias{targomoDependency}
\title{Dependencies}
\usage{
targomoDependency(...)
}
\arguments{
\item{...}{Further args to pass to \code{htmltools::htmlDependency}}
}
\description{
This function exposes the bundled API prod files and bindings as dependencies.
}
