% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{deriveSources}
\alias{deriveSources}
\alias{createIds}
\alias{createPoints}
\alias{deriveTargets}
\title{Derive Sources/Targets}
\usage{
createIds(data = NULL, id = NULL)

createPoints(data = NULL, lat = NULL, lng = NULL, id = NULL)

deriveSources(points, options)

deriveTargets(points)
}
\arguments{
\item{data}{The data object}

\item{id}{The id vector or formula to resolve}

\item{lat, lng}{The lat/lng vectors or formulae to resolve}

\item{points}{A processed data object (for sources/targets).}

\item{options}{A processed options object (for sources).}
}
\value{
A data.frame of sources/targets, with IDs attached.
}
\description{
Function to create the sources needed to query the Targomo API.
}
