% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn-drawing.R
\name{draw}
\alias{draw}
\alias{drawRoute}
\alias{drawWalk}
\alias{drawBike}
\alias{drawCar}
\alias{drawTransit}
\title{Draw Routes}
\usage{
drawRoute(map, route, drawOptions, type, ...)

drawWalk(map, route, drawOptions, ...)

drawBike(map, route, drawOptions, ...)

drawCar(map, route, drawOptions, ...)

drawTransit(map, route, drawOptions, ...)
}
\arguments{
\item{map}{A leaflet map.}

\item{route}{A route object to draw.}

\item{drawOptions}{Drawing options provided by \code{\link{routeDrawOptions}}.}

\item{type}{What route type to draw.}

\item{...}{Further arguments to pass to leaflet functions.}
}
\description{
Helper functions for drawing different routes.
}
