% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoPolygons.R
\name{getTargomoGeoJSON}
\alias{getTargomoGeoJSON}
\title{TODO: add function to create a source object, so that it can be mapped over
the rows of the output of leaflet::derivePoints.
Get Targomo Polygon GeoJSON data}
\usage{
getTargomoGeoJSON(api_key = Sys.getenv("TARGOMO_API_KEY"),
  region = Sys.getenv("TARGOMO_REGION"), data = NULL, lat = NULL,
  lng = NULL, options = targomoOptions(), ...)
}
\arguments{
\item{...}{}
}
\description{
Function to call the Targomo API and return geojson.
}
