% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{getTargomoMapURL}
\alias{getTargomoMapURL}
\title{Targomo Map Tiles URL}
\usage{
getTargomoMapURL(style = "basic", api_key = Sys.getenv("TARGOMO_API_KEY"))
}
\arguments{
\item{style}{A valid Targomo Map Style - see \code{targomoMapStyles()}}

\item{api_key}{Your Targomo API key - defaults to the \code{TARGOMO_API_KEY}
ennvironment variable}
}
\value{
The URL of the requested map tile
}
\description{
Targomo Map Tiles URL
}
\examples{
getTargomoMapURL(style = "toner", api_key = NULL)

}
