% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoPolygons.R
\name{getTargomoPolygons}
\alias{getTargomoPolygons}
\title{Get Targomo Polygon GeoJSON data}
\usage{
getTargomoPolygons(api_key = Sys.getenv("TARGOMO_API_KEY"),
  region = Sys.getenv("TARGOMO_REGION"), data = NULL, lat = NULL,
  lng = NULL, options = targomoOptions(), ...)
}
\arguments{
\item{...}{}
}
\description{
Function to call the Targomo API and return geojson.
}
