% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{messageMultipleTravelModes}
\alias{messageMultipleTravelModes}
\title{Message if multiple Travel Modes supplied}
\usage{
messageMultipleTravelModes(tms)
}
\arguments{
\item{tms}{A vector of travel modes}
}
\description{
Message if multiple Travel Modes supplied
}
