% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\alias{targomoOptions}
\title{Set Targomo Options}
\usage{
targomoOptions(travelType = "bike", travelTimes = list(600, 1200,
  1800), intersectionMode = "union", carRushHour = FALSE,
  walkSpeed = 5, walkUpHillAdjustment = 10,
  walkDownHillAdjustment = 0, bikeSpeed = 15,
  bikeUpHillAdjustment = 20, bikeDownHillAdjustment = -10,
  transitDate = NULL, transitTime = NULL, transitDuration = NULL,
  transitMaxWalkingTimeFromSource = NULL,
  transitMaxWalkingTimeToTarget = NULL, transitEarliestArrival = FALSE,
  transitMaxTransfers = NULL, edgeWeight = "time",
  maxEdgeWeight = 1800, elevation = FALSE, serializer = "geojson",
  srid = 4326, minPolygonHoleSize = NULL, buffer = NULL,
  simplify = NULL, quadrantSegments = NULL, decimalPrecision = NULL)
}
\arguments{
\item{travelType}{What mode of transport to use - car, bike, walk or public transport.}

\item{travelTimes}{A list of times - each time corresponds to a
different polygon. Your API key will determine how many you can add. Acceptable formats
are numeric (interpreted as seconds), or a string of the form \code{.h.m.s}. E.g. for
1 hour, "1h", 90 minutes = "1h30m" or "90m" etc.}

\item{intersectionMode}{Whether to calculate the union or intersection of multiple sources.}

\item{carRushHour}{Account for rush hour while driving.}

\item{walkSpeed, walkUpHillAdjustment, walkDownHillAdjustment}{Settings for walking travel type.}

\item{bikeSpeed, bikeUpHillAdjustment, bikeDownHillAdjustment}{Settings for cycling travel type.}

\item{transitDate}{The date for public transport calculations (YYYYMMDD).}

\item{transitTime}{The time in seconds since midnight to begin transit.}

\item{transitDuration}{The duration of the transit timeframe (seconds or \code{.h.m.s} string).}

\item{transitMaxWalkingTimeFromSource, transitMaxWalkingTimeToTarget}{Settings
for transit travel type.}

\item{transitMaxTransfers, transitEarliestArrival}{Further transit settings.}

\item{edgeWeight}{Should calculations be in "time" or "distance"?}

\item{maxEdgeWeight}{The max time or distance to search for routes/times in - acceptable
formats are numeric (seconds or metres), or a string of the form \code{.h.m.s} for time,
or \code{.km.m.ml} for distance, where ml represents miles (1609 metres).}

\item{elevation}{Account for elevation?}

\item{serializer}{Should be "geojson" or "json". See API for details.}

\item{srid}{The spatial reference of the returned data.}

\item{minPolygonHoleSize}{Minimum area of holes in returned polygons.}

\item{simplify, buffer}{Parameters for manipulating the returned polygons.}

\item{quadrantSegments, decimalPrecision}{Parameters for fine-tuning the returned polygons.}
}
\value{
A (filtered, no NULLs) list of options to pass to the API
}
\description{
This function sets the options to be passed to the API service.
For full details of available options see:
\url{https://docs.targomo.com/core/}
}
\examples{
targomoOptions(travelType = "transit")
targomoOptions(travelTimes = list("15m", "30m", "45m", "1h"), maxEdgeWeight = "1h")

}
