% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoPointsofInterest.R
\name{points-of-interest}
\alias{points-of-interest}
\alias{translatePOIs}
\alias{getTargomoPOIs}
\title{Add Targomo Points of Interest to a Leaflet Map}
\usage{
translatePOIs(keyvalues)

getTargomoPOIs(source_data = NULL, source_lat = NULL,
  source_lng = NULL, options = targomoOptions(),
  api_key = Sys.getenv("TARGOMO_API_KEY"),
  region = Sys.getenv("TARGOMO_REGION"), config = list(),
  verbose = FALSE, progress = FALSE)
}
\arguments{
\item{source_data}{The data object from which source ppoints are derived.}

\item{source_lng, source_lat}{Vectors/one-sided formulas of longitude and latitude.}

\item{options}{A list of \code{\link{targomoOptions}} to call the API.}

\item{api_key}{Your Targomo API key - defaults to the \code{TARGOMO_API_KEY}
ennvironment variable}

\item{region}{Your Targomo region - defaults to the \code{TARGOMO_REGION}
environment variable}

\item{config}{Config options to pass to \code{httr::POST} e.g. proxy settings}

\item{verbose}{Whether to print out information about the API call.}

\item{progress}{Whether to show a progress bar of the API call.}

\item{map}{A leaflet map.}

\item{group}{The leaflet map group to add the points to. A single group is used
for all the points added by one API call.}
}
\description{
Function for adding targomo points of interest to a leaflet map.
}
