% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities.R
\name{print.tgm_capabilities}
\alias{print.tgm_capabilities}
\title{Print Method for Capabilities}
\usage{
\method{print}{tgm_capabilities}(caps)
}
\arguments{
\item{caps}{A list, output of \code{\link{getTargomoCapabilities}}}
}
\value{
The list, invisibly
}
\description{
A custom print method to make the capabilities list appear nicely in the console.
}
