% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn-processing.R
\name{process}
\alias{process}
\alias{catchBadResponse}
\alias{processResponse}
\alias{processPolygons}
\alias{getRouteFeature}
\alias{processRoutes}
\alias{processTimes}
\title{Process API responses}
\usage{
catchBadResponse(response)

processResponse(response, service)

processPolygons(payload)

getRouteFeature(route, feature)

processRoutes(payload)

processTimes(payload)
}
\arguments{
\item{response}{A response object from \code{\link{callTargomoAPI}}.}

\item{service}{The Targomo API service being called - polygon, route or time.}

\item{payload}{The \code{httr::content} of the response.}

\item{route}{A single element of the returned routes list.}

\item{feature}{A route feature to extract (e.g. car sections, bike sections...)}
}
\description{
Functions to turn a successful request into data - either polygons, routes or times.
}
