% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities.R
\name{processCapabilities}
\alias{processCapabilities}
\title{Process Capabilities}
\usage{
processCapabilities(response)
}
\arguments{
\item{response}{The API response object}
}
\value{
The formatted list, of class 'tgm_capabilitites'
}
\description{
This function takes the raw JSON list of capabilities and converts then into a formatted list.
}
