% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn-polygon.R
\name{processPolygonResponse}
\alias{processPolygonResponse}
\title{Create Targomo Polygons}
\usage{
processPolygonResponse(response, output = "sf")
}
\arguments{
\item{response}{A response object from \code{\link{callPolygonService}}.}

\item{output}{Either 'sf' or 'geojson'.}
}
\description{
Function to turn a successful request into polygons.
}
