% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn-processing.R
\name{processResponse}
\alias{processResponse}
\alias{catchBadResponse}
\alias{processPolygons}
\alias{processRoutes}
\alias{processTime}
\title{Process API responses}
\usage{
catchBadResponse(response)

processResponse(response, service)

processPolygons(payload)

processRoutes(payload)

processTime(payload)
}
\arguments{
\item{response}{A response object from \code{\link{callPolygonService}}.}

\item{service}{The Targomo API service being called - polygon, route or time.}
}
\description{
Functions to turn a successful request into data - either polygons, routes or times.
}
