% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setTargomoAPIKey}
\alias{setTargomoAPIKey}
\title{Set Targomo API Key}
\usage{
setTargomoAPIKey(api_key, overwrite = FALSE)
}
\arguments{
\item{api_key}{Your Targomo API key}

\item{overwrite}{Whether to overwrite an existing setting}
}
\description{
This helper function allows you to set your API key in either a global or local
.Renviron file, for ease of use. All of the \code{TargomoR} functions which call
the Targomo API require an API key, and use the \code{TARGOMO_API_KEY} environment
variable by default.
}
