% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoPolygons.R
\name{targomo-options}
\alias{targomo-options}
\alias{targomoOptions}
\title{Set Targomo Options}
\usage{
targomoOptions(travelTimes = c(600, 1200, 1800), travelType = "bike",
  strokeWidth = 20, inverse = FALSE)
}
\arguments{
\item{travelTimes}{A vector of times, in seconds - each time corresponds to a
different polygon. Your API key will determine how many you can add.}

\item{travelType}{What mode of transport to use - car, bike, walk or public transport.}

\item{strokeWidth}{The weight of stroke used to draw the polygons.}

\item{inverse}{Should the polygons be inverted?}
}
\description{
This function sets the options to be passed to the API service.
}
