% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoDependency.R
\name{targomoDependency}
\alias{targomoDependency}
\title{Function to create dependencies for each method}
\usage{
targomoDependency(...)
}
\description{
Function to create dependencies for each method
}
