% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoPolygons.R
\name{targomoOptions}
\alias{targomoOptions}
\title{Set Targomo Options}
\usage{
targomoOptions(travelTimes = c(600, 1200, 1800), travelType = "car",
  stroke = 10, invert = FALSE)
}
\description{
Set Targomo Options
}
