% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities.R
\name{tidy-capabilities}
\alias{tidy-capabilities}
\alias{tidyGeneral}
\alias{tidyTransit}
\alias{tidySpeeds}
\alias{print.tgm_capabilities}
\title{Helper functions for tidying up capabilities response}
\usage{
tidyGeneral(general)

tidyTransit(transit)

tidySpeeds(speeds)

\method{print}{tgm_capabilities}(caps)
}
\arguments{
\item{general, transit, speeds}{The parts of the list}

\item{caps}{A capabilities list from \code{\link{getTargomoCapabilities}}}
}
\description{
Helper functions for tidying up capabilities response
}
