% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoTimes.R
\name{times}
\alias{times}
\alias{getTargomoTimes}
\alias{addTargomoTimes}
\title{Add Targomo Times to a Leaflet Map}
\usage{
getTargomoTimes(source_data = NULL, source_lat = NULL,
  source_lng = NULL, target_data = NULL, target_lat = NULL,
  target_lng = NULL, options = targomoOptions(),
  api_key = Sys.getenv("TARGOMO_API_KEY"),
  region = Sys.getenv("TARGOMO_REGION"), verbose = FALSE,
  progress = FALSE)

addTargomoTimes(map, source_data = NULL, source_lat = NULL,
  source_lng = NULL, target_data = NULL, target_lat = NULL,
  target_lng = NULL, options = targomoOptions(),
  drawOptions = timeDrawOptions(), group = NULL,
  api_key = Sys.getenv("TARGOMO_API_KEY"),
  region = Sys.getenv("TARGOMO_REGION"), verbose = FALSE,
  progress = FALSE)
}
\arguments{
\item{source_data, target_data}{The source and target points for your travel times -
supported types are data.frame matrix and objects from the sf and sp packages.}

\item{source_lat, source_lng}{Columns identifying the latitude
and longitude columns in your source data, or numeric vectors of equal length.}

\item{target_lat, target_lng}{As for \code{source_lat,source_lng} but for target data.}

\item{options}{A list of \code{\link{targomoOptions}} to send to the API.}

\item{api_key}{Your Targomo API key - defaults to the \code{TARGOMO_API_KEY}
ennvironment variable.}

\item{region}{Your Targomo region - defaults to the \code{TARGOMO_REGION}
environment variable.}

\item{verbose}{Whether to print out information about the API call.}

\item{progress}{Whether to show a progress bar of the API call.}

\item{map}{A leaflet map}

\item{drawOptions}{A list of \code{\link{timeDrawOptions}} to determine how to show
the resulting times on the map.}

\item{group}{The leaflet map group to add the times to. One group is used for all
map elements being drawn per call to the API.}
}
\description{
This function takes source and target data, together with options for the API and
drawing options, and returns the map with the requested travel time data.
}
