% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoPolygons.R
\name{addTargomoPolygons}
\alias{addTargomoPolygons}
\title{Add Targomo Polygons}
\usage{
addTargomoPolygons(map, api_key = Sys.getenv("TARGOMO_API_KEY"),
  region = Sys.getenv("TARGOMO_REGION"), lat = 55.9, lng = -3.1,
  options = targomoOptions(), layerId = NULL, group = NULL,
  fitBounds = TRUE)
}
\arguments{
\item{map}{A leaflet map}

\item{api_key}{Your Targomo API key - defaults to the \code{TARGOMO_API_KEY} environment variable.}

\item{region}{Your Targomo region - defaults to the \code{TARGOMO_REGION} environment variable.
See here for the different regions: \url{https://targomo.com/developers/availability/}}

\item{lat, lng}{Coordinates of the source of the polygons}

\item{options}{See \code{\link{targomoOptions}}.}

\item{layerId}{The leaflet map layerId for the resulting polygons.}

\item{group}{The leaflet map group for the polygons.}

\item{fitBounds}{Should the map rezoom to fit the bounds of the polygons?}
}
\description{
This function and the associated options allow you to easily add isochrones from
the Targomo Polygon API to your leaflet map.
}
