% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoPolygons.R
\name{getTargomoPolygons}
\alias{getTargomoPolygons}
\alias{addTargomoPolygons}
\title{Get Targomo Polygons}
\usage{
getTargomoPolygons(api_key = Sys.getenv("TARGOMO_API_KEY"),
  region = Sys.getenv("TARGOMO_REGION"), data = NULL, lat = NULL,
  lng = NULL, options = targomoOptions(), sf = TRUE,
  verbose = FALSE, progress = FALSE)

addTargomoPolygons(map, api_key = Sys.getenv("TARGOMO_API_KEY"),
  region = Sys.getenv("TARGOMO_REGION"), lng = NULL, lat = NULL,
  options = targomoOptions(), layerId = NULL, group = NULL,
  stroke = TRUE, weight = 5, color = "#03F", opacity = 0.5,
  fill = TRUE, fillColor = color, fillOpacity = 0.2,
  dashArray = NULL, smoothFactor = 1, noClip = FALSE, popup = NULL,
  popupOptions = NULL, label = NULL, labelOptions = NULL,
  pathOptions = leaflet::pathOptions(), highlightOptions = NULL,
  data = leaflet::getMapData(map), ...)
}
\arguments{
\item{api_key}{Your Targomo API key - defaults to the \code{TARGOMO_API_KEY} env var.}

\item{region}{Your Targomo region - defaults to the \code{TARGOMO_REGION} env var.}

\item{data}{The data object from which arguments are derived.}

\item{lng, lat}{Vectors/one-sided formulas of longitude and latitude.}

\item{options}{A list of \code{\link{targomoOptions}} to call the API.}

\item{sf}{Logical - return an sf object?}

\item{verbose}{Logical - show the API call?}

\item{progress}{Logical - show a progress bar for the API call?}

\item{map}{A leaflet map.}

\item{layerId}{The layer ID.}

\item{group}{The layer group.}

\item{stroke}{Should the borders be drawn?}

\item{weight}{Stroke width in pixels.}

\item{color}{Stroke color.}

\item{opacity}{Stroke opacity.}

\item{fill}{Should the polygons be filled with colour?}

\item{fillColor}{The fill colour.}

\item{fillOpacity}{The fill opacity.}

\item{dashArray}{A string to define the stroke dash pattern.}

\item{smoothFactor}{How much to simplify polylines on each zoom level.}

\item{noClip}{Whether to disable polyline clipping.}

\item{popup}{HTML content for the popup.}

\item{popupOptions}{A list of \code{\link[leaflet]{popupOptions}}.}

\item{label}{A character vector of labels.}

\item{labelOptions}{A list of \code{\link[leaflet]{labelOptions}}.}

\item{pathOptions}{Options for the path.}

\item{highlightOptions}{A list of \code{\link[leaflet]{highlightOptions}}.}

\item{...}{Further options to pass to \code{\link{getTargomoPolygons}} e.g. verbose.}
}
\description{
Function for adding targomo polygons to a leaflet map - uses \link[leaflet]{addPolygons}
as a workhorse function, and adds the polygons as an sf object.
}
