% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoPolygons.R
\name{addTargomoPolygons}
\alias{addTargomoPolygons}
\title{Add Targomo Polygons}
\usage{
addTargomoPolygons(map, api_key = Sys.getenv("TARGOMO_API_KEY"),
  lng = 0, lat = 51, options = targomoOptions())
}
\description{
Add Targomo Polygons
}
