% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn-API.R
\name{callTargomoAPI}
\alias{callTargomoAPI}
\title{Call the Targomo API}
\usage{
callTargomoAPI(api_key = Sys.getenv("TARGOMO_API_KEY"),
  region = Sys.getenv("TARGOMO_REGION"), service, sources,
  targets = NULL, options, verbose = FALSE, progress = FALSE)
}
\arguments{
\item{api_key}{The Targomo API key.}

\item{region}{The Targomo region.}

\item{service}{The Targomo service - 'polygon', 'route', or 'time'.}

\item{sources}{A processed sources object to pass to the API.}

\item{targets}{A processed targets object (optional).}

\item{options}{A processed options list.}

\item{verbose}{Display info on the API call?}

\item{progress}{Display a progress bar?}
}
\description{
Call the Targomo API
}
