% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn-API.R
\name{createRequestBody}
\alias{createRequestBody}
\title{Create Request Body}
\usage{
createRequestBody(service, sources = NULL, targets = NULL, options)
}
\arguments{
\item{service}{The Targomo Service to create a body for - 'polygon', 'time', 'route'.}

\item{sources}{A processed sources object to pass to the API.}

\item{targets}{A processed targets object (optional).}

\item{options}{A processed options list.}
}
\description{
Function to create a request body using the sources and options given.
}
