% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn-API.R
\name{deriveSources}
\alias{deriveSources}
\alias{deriveIds}
\alias{deriveTargomoPoints}
\alias{deriveTargets}
\title{Derive Sources/Targets}
\usage{
deriveIds(data = NULL, id = NULL)

deriveTargomoPoints(data = NULL, lat = NULL, lng = NULL, id = NULL)

deriveSources(data = NULL, lat = NULL, lng = NULL, id = NULL,
  options)

deriveTargets(data = NULL, lat = NULL, lng = NULL, id = NULL)
}
\arguments{
\item{data}{The data object}

\item{id}{The id vector or formula to resolve}

\item{lat, lng}{The lat/lng vectors or formulae to resolve}

\item{options}{A processed options object (for sources).}
}
\description{
Function to create the sources needed to query the Targomo API.
}
