% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn-API.R
\name{deriveSources}
\alias{deriveSources}
\alias{deriveTargets}
\title{Derive Sources/Targets}
\usage{
deriveSources(data, lat = NULL, lng = NULL, options)

deriveTargets(data, lat = NULL, lng = NULL)
}
\arguments{
\item{data}{The data object}

\item{lat, lng}{The lat/lng vectors or formulae to resolve}

\item{options}{A processed options object (for sources).}
}
\description{
Function to create the sources needed to query the Targomo API.
}
