% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn-drawing.R
\name{draw}
\alias{draw}
\alias{drawRouteSegment}
\alias{drawWalk}
\alias{drawBike}
\alias{drawCar}
\alias{drawTransit}
\alias{drawRoutes}
\title{Draw Routes}
\usage{
drawRouteSegment(map, segment, drawOptions, type, ...)

drawWalk(map, segment, drawOptions, ...)

drawBike(map, segment, drawOptions, ...)

drawCar(map, segment, drawOptions, ...)

drawTransit(map, segment, drawOptions, ...)

drawRoutes(map, routes, drawOptions = routeDrawOptions(), group = NULL)
}
\arguments{
\item{map}{A leaflet map.}

\item{segment}{A route segment object to draw.}

\item{drawOptions}{Drawing options provided by \code{\link{routeDrawOptions}}.}

\item{type}{What route type to draw.}

\item{...}{Further arguments to pass to leaflet functions.}

\item{routes}{A list of route segments provided by \code{\link{getTargomoRoutes}}.}
}
\description{
Helper functions for drawing different routes.
}
