% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{draw}
\alias{draw}
\alias{drawRouteSegment}
\alias{drawWalk}
\alias{drawBike}
\alias{drawCar}
\alias{drawTransit}
\alias{drawRoutes}
\title{Draw Routes}
\usage{
drawRouteSegment(map, segment, drawOptions, type, group, ...)

drawWalk(map, segment, drawOptions, group, ...)

drawBike(map, segment, drawOptions, group, ...)

drawCar(map, segment, drawOptions, group, ...)

drawTransit(map, segment, drawOptions, group, ...)

drawRoutes(map, routes, drawOptions = routeDrawOptions(), group = NULL,
  ...)
}
\arguments{
\item{map}{A leaflet map.}

\item{segment}{A route segment object to draw.}

\item{drawOptions}{Drawing options provided by \code{\link{routeDrawOptions}}.}

\item{type}{What route type to draw.}

\item{group}{The leaflet map group to add the routes to.}

\item{...}{Further arguments to pass to leaflet functions.}

\item{routes}{A list of route segments provided by \code{\link{getTargomoRoutes}}.}
}
\description{
Helper functions for drawing different routes.
}
