% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn-formatting.R
\name{formatting}
\alias{formatting}
\alias{formatEdgeWeight}
\title{Format Requests and Responses}
\usage{
formatEdgeWeight(edgeWeight, type)
}
\arguments{
\item{edgeWeight}{A time or distance, in numeric or string form.}

\item{type}{Either 'time' or 'distance'.}
}
\description{
Functions to make the interface easier and more intuitive to use.
}
