% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{formatting}
\alias{formatting}
\alias{numericEdgeWeight}
\alias{prettyEdgeWeight}
\title{Format Requests and Responses}
\usage{
numericEdgeWeight(edgeWeight, type)

prettyEdgeWeight(edgeWeight, type)
}
\arguments{
\item{edgeWeight}{A time or distance, in numeric or string form.}

\item{type}{Either 'time' or 'distance'.}
}
\description{
Functions to make the interface easier and more intuitive to use.
}
