% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targomoPolygons.R
\name{targomoOptions}
\alias{targomoOptions}
\title{Set Targomo Options}
\usage{
targomoOptions(travelTimes = c(600, 1200, 1800), travelType = "bike",
  intersectionMode = "union", serializer = "geojson", srid = 4326,
  ...)
}
\arguments{
\item{travelTimes}{A vector of times, in seconds - each time corresponds to a
different polygon. Your API key will determine how many you can add.}

\item{travelType}{What mode of transport to use - car, bike, walk or public transport.}

\item{intersectionMode}{Whether to calculate the union or intersection of multiple sources}

\item{...}{Further arguments to pass to the API - see \url{https://docs.targomo.com/core/#/Polygon_Service/post_westcentraleurope_v1_polygon}}
}
\description{
This function sets the options to be passed to the API service.
}
