% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities.R
\name{tidy-capabilities}
\alias{tidy-capabilities}
\alias{tidyGeneral}
\alias{tidyTransit}
\alias{tidySpeeds}
\title{Helper functions for tidying up capabilities response}
\usage{
tidyGeneral(general)

tidyTransit(transit)

tidySpeeds(speeds)
}
\arguments{
\item{general, transit, speeds}{The parts of the list}
}
\value{
Lists of capabilities (data.frames or vectors)
}
\description{
These functions tidy up the raw capabilities lists.
}
