% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\name{timeLegendOptions}
\alias{timeLegendOptions}
\title{Time Legend Options}
\usage{
timeLegendOptions(
  position = "topright",
  title = "Travel Times",
  layerId = NULL
)
}
\arguments{
\item{position}{One of c("topright", "topleft", "bottomright", "bottomleft").}

\item{title}{The legend title.}

\item{layerId}{The legend layer ID.}
}
\value{
A list of options governing how the time legend appears on the map
}
\description{
Time Legend Options
}
\examples{
# show the list
timeLegendOptions()

}
